requirements.txt
fastapi
pinecone
pydantic
python-dotenv
uvicorn
tiktoken
langchain
langchain-openai
langchain-pinecone
langchain-text-splitters